services/grpc-suite/requirements.txt
grpcio==1.63.0
grpcio-tools==1.63.0
grpcio-reflection==1.63.0
protobuf==4.25.3

# PostgreSQL
asyncpg==0.29.0

# MongoDB
motor==3.4.0

# Redis
redis[asyncio]==5.0.3

# Observability (Prometheus metrics exposed on a sidecar HTTP port)
prometheus-client==0.20.0

# Utilities
python-dotenv==1.0.1
